BlackPoint & Camera WB GUI Sync Issue
Set "Black Point To 0" and "Use Camera White Balance" settings were only set in the raw processor when ticked. When subsequently unticked, their settings values were not set and thus never applied correctly. Whilst auto WB is not used, it would have had this same behaviour too, so a fix for that has been made too for if/when it is.