Add /DSTRIP_SIZE_DEFAULT=262144 to compile options for libtiff.
The default is 8192.   This allows TIFFReadEncodedStrip to read as many rows as will fit into 256kB (assuming that Rows/Strip in the TIFF Directory is set large enough).   For a TIFF file created by a Nikon D4s which set Rows/Strip to 3820 (number of rows in the image), and a row size of 14784, this allows libtiff to read 17 rows in one call:

>TIFF Scan Line Size 14784
>TIFF spp=3, bpp=8, w=4928, h=3280
>Allocating buffer of 48491520 bytes
>Number of strips is 193

meaning 193 calls to TIFFReadEncodedStrip instead of 3820 calls.