DSSCL Bugfix: Register Only Was Not Working.
Error in the logic meant you had to both register and stack - it should allow one or the other (or both). Tidied up help formatting and added verbose error if file list is invalid. Also skip 1st param as it was being interpreted as the file list if no other file list was passed.