Implement atexitHandler for resource cleanup
Added an `atexitHandler` function in `DeepSkyStacker.cpp`, `DeepSkyStackerCL.cpp`, and `DeepSkyStackerLive.cpp` to manage trace file cleanup on program termination. Registered this handler in the `main` function using `std::atexit()`.   Moved code to remove or retain the trace file to the new `atexitHandler`. Updated the `terminate` method in `tracecontrol.h` to remove call to ZTrace::disableTrace().